ml/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
opencv-python==4.8.1.78
pytesseract==0.3.10
numpy==1.26.2
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
reportlab==4.0.8
PyPDF2==3.0.1

# Machine Learning packages
pandas==2.1.3
scikit-learn==1.3.2
joblib==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
